refactor: general improvements
